Move repositories to separate sub-module
