+++ src/conduit/articles_repository.rs
@@ -2,13 +2,14 @@ use crate::conduit::{articles, favorites, followers, users};
 use crate::db::models::{Article, NewArticle};
 use crate::db::Repo;
 use crate::domain;
-use crate::domain::{GetUserError, UsersRepository};
+use crate::domain::repositories::{ArticleRepository, UsersRepository};
+use crate::domain::GetUserError;
 use diesel::PgConnection;
 use uuid::Uuid;
 
 pub struct Repository<'a>(pub &'a Repo<PgConnection>);
 
-impl<'a> domain::ArticleRepository for Repository<'a> {
+impl<'a> ArticleRepository for Repository<'a> {
     fn publish(
         &self,
         draft: domain::ArticleDraft,
@@ -63,7 +64,7 @@ impl<'a> domain::ArticleRepository for Repository<'a> {
     }
 }
 
-impl<'a> domain::UsersRepository for Repository<'a> {
+impl<'a> UsersRepository for Repository<'a> {
     fn get_by_id(&self, user_id: Uuid) -> Result<domain::User, GetUserError> {
         let u = users::find(&self.0, user_id)?;
         let profile = domain::Profile::new(u.username, u.bio, u.image);
